LPD-82692 Transfer content page integration tests
