+++ modules/test/playwright/tests/object-web/main/objectDefinition.spec.ts
@@ -11,7 +11,6 @@ import {
 } from '@liferay/object-admin-rest-client-js';
 import {expect, mergeTests} from '@playwright/test';
 
-import {collectionsPagesTest} from '../../../fixtures/collectionsPagesTest';
 import {dataApiHelpersTest} from '../../../fixtures/dataApiHelpersTest';
 import {displayPageTemplatesPagesTest} from '../../../fixtures/displayPageTemplatesPagesTest';
 import {featureFlagsTest} from '../../../fixtures/featureFlagsTest';
@@ -25,13 +24,11 @@ import getRandomString from '../../../utils/getRandomString';
 import {performUserSwitch, userData} from '../../../utils/performLogin';
 import {waitForAlert} from '../../../utils/waitForAlert';
 import getFormContainerDefinition from '../../layout-content-page-editor-web/main/utils/getFormContainerDefinition';
-import getFragmentDefinition from '../../layout-content-page-editor-web/main/utils/getFragmentDefinition';
 import getPageDefinition from '../../layout-content-page-editor-web/main/utils/getPageDefinition';
 import {localizationPagesTest} from '../../site-admin-web/main/fixtures/localizationPagesTest';
 import {generateObjectFields} from '../utils/generateObjectFields';
 
 const test = mergeTests(
-	collectionsPagesTest,
 	dataApiHelpersTest,
 	displayPageTemplatesPagesTest,
 	featureFlagsTest({
@@ -1651,144 +1648,6 @@ test.describe('Manage object definitions through View Object Definitions', () =>
 
 		await expect(page.getByText('Required')).toBeVisible();
 	});
-});
-
-test.describe('Manage object definitions through a Page', () => {
-	test('can display an object reactivated on the Collection Providers', async ({
-		apiHelpers,
-		collectionsPage,
-		page,
-		site,
-		viewObjectDefinitionsPage,
-	}) => {
-		const objectDefinition =
-			await apiHelpers.objectAdmin.postRandomObjectDefinition({
-				status: {code: 0},
-			});
-
-		apiHelpers.data.push({
-			id: objectDefinition.id,
-			type: 'objectDefinition',
-		});
-
-		await viewObjectDefinitionsPage.goto();
-
-		await viewObjectDefinitionsPage.changeObjectActivateStatus(
-			objectDefinition.name
-		);
-
-		await viewObjectDefinitionsPage.goto();
-
-		await viewObjectDefinitionsPage.changeObjectActivateStatus(
-			objectDefinition.name
-		);
-
-		await collectionsPage.goto(site.friendlyUrlPath);
-
-		await page.getByRole('link', {name: 'Collection Providers'}).click();
-
-		await expect(
-			page.getByText(objectDefinition.name).first()
-		).toBeVisible();
-	});
-
-	test('can display an object reactivated on the Page Item Selector', async ({
-		apiHelpers,
-		page,
-		pageEditorPage,
-		site,
-		viewObjectDefinitionsPage,
-	}) => {
-		const objectDefinition =
-			await apiHelpers.objectAdmin.postRandomObjectDefinition({
-				status: {code: 0},
-			});
-
-		apiHelpers.data.push({
-			id: objectDefinition.id,
-			type: 'objectDefinition',
-		});
-
-		await viewObjectDefinitionsPage.goto();
-
-		await viewObjectDefinitionsPage.changeObjectActivateStatus(
-			objectDefinition.name
-		);
-
-		await viewObjectDefinitionsPage.goto();
-
-		await viewObjectDefinitionsPage.changeObjectActivateStatus(
-			objectDefinition.name
-		);
-
-		const headingDefinition = getFragmentDefinition({
-			id: getRandomString(),
-			key: 'BASIC_COMPONENT-heading',
-		});
-
-		const layout = await apiHelpers.headlessDelivery.createSitePage({
-			pageDefinition: getPageDefinition([headingDefinition]),
-			siteId: site.id,
-			title: getRandomString(),
-		});
-
-		await pageEditorPage.goto(layout, site.friendlyUrlPath);
-
-		await page.getByText('Heading Example', {exact: true}).dblclick();
-
-		await page.getByLabel('Select Item').click();
-
-		await expect(
-			page
-				.frameLocator('iframe[title="Select"]')
-				.getByRole('menuitem', {name: objectDefinition.name})
-		).toBeVisible();
-	});
-
-	test(
-		'Cannot select unpublished object for a display page template',
-		{tag: '@LPS-137871'},
-		async ({apiHelpers, displayPageTemplatesPage, page}) => {
-			const objectName = 'CustomObject' + getRandomInt();
-
-			const objectDefinition =
-				await apiHelpers.objectAdmin.postRandomObjectDefinition({
-					objectDefinitionExternalReferenceCode: objectName,
-					status: {code: 2},
-				});
-
-			apiHelpers.data.push({
-				id: objectDefinition.id,
-				type: 'objectDefinition',
-			});
-
-			await test.step('Create a new blank display page template', async () => {
-				await displayPageTemplatesPage.goto();
-
-				await page.getByRole('button', {name: 'New'}).click();
-
-				await page
-					.getByRole('menuitem', {name: 'Display Page Template'})
-					.click();
-
-				await page.getByRole('button', {name: 'Blank'}).click();
-			});
-
-			await test.step('Verify unpublished object is not in content type options', async () => {
-				const contentTypeSelect = page.getByLabel('Content Type');
-
-				await expect(contentTypeSelect).toBeVisible();
-
-				const options = contentTypeSelect.locator('option');
-
-				const optionTexts = await options.allTextContents();
-
-				expect(optionTexts).not.toContain(
-					objectDefinition.label['en_US']
-				);
-			});
-		}
-	);
 });
 
 cmsTest.describe('Manage enableFormContainer configuration', () => {
